# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Presentation Studio API",
        "version": "1.0.0",
        "description": "API for Presentation Studio. Admin endpoints require a 94-character API token in the Authorization header."
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "TOKEN",
                "description": "Enter your 94-character API token (only for admin endpoints)"
            }
        }
    },
    "paths": {
...
        "/api/project/{name}": {
            "get": {
                "summary": "Get project data",
                "description": "Returns data for a specific project. No token required.",
                "tags": [
                    "Projects"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Project name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            }
        },
...
        "/api/admin/projects": {
            "get": {
                "summary": "Get all projects",
                "description": "Returns all projects from all users. **Requires 94-character API token**.",
                "tags": [
                    "Projects (Admin)"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token"
                    }
                }
            }
        },
...
    }
}